Added params to some endpoints + changed PersonalId to long

+++ MedicationPrescriber.Api/Dtos/PatientDto.cs
@@ -5,7 +5,7 @@ namespace MedicationPrescriber.Api.Dtos
 {
     public class PatientDto
     {
-        public int PersonalId { get; set; }
+        public long PersonalId { get; set; }
 
         public DateTime Birthdate { get; set; }
